# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Pedidos - Desafio Técnico",
    "description": "API para gerenciamento de pedidos com integração MongoDB e transformação de dados.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor Local"
    }
  ],
  "paths": {
    "/order": {
      "post": {
        "summary": "Criar um novo pedido",
        "description": "Recebe os dados em Português e salva no banco em Inglês (Mapping).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PedidoInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pedido criado com sucesso"
          },
          "400": {
            "description": "Dados inválidos"
          }
        }
      }
    },
    "/order/list": {
      "get": {
        "summary": "Listar todos os pedidos",
        "responses": {
          "200": {
            "description": "Lista de pedidos retornada com sucesso"
          }
        }
      }
    },
    "/order/{id}": {
      "get": {
        "summary": "Buscar pedido por ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "string" },
            "required": true,
            "description": "ID do pedido (ex: v10089015vdb-01)"
          }
        ],
        "responses": {
          "200": { "description": "Pedido encontrado" },
          "404": { "description": "Pedido não encontrado" }
        }
      },
      "put": {
        "summary": "Atualizar um pedido",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PedidoInput" }
            }
          }
        },
        "responses": {
          "200": { "description": "Atualizado com sucesso" }
        }
      },
      "delete": {
        "summary": "Deletar um pedido",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Deletado com sucesso" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PedidoInput": {
        "type": "object",
        "properties": {
          "numeroPedido": { "type": "string", "example": "v10089015vdb-01" },
          "valorTotal": { "type": "number", "example": 10000 },
          "dataCriacao": { "type": "string", "format": "date-time", "example": "2023-07-19T12:24:11.529Z" },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "idItem": { "type": "string", "example": "2434" },
                "quantidadeItem": { "type": "number", "example": 1 },
                "valorItem": { "type": "number", "example": 1000 }
              }
            }
          }
        }
      }
    }
  }
}